refactor - root 로만 쓰던 `displayChilds` 기능 공통사용
