--- src/utils/displayRoot.js
@@ -1,11 +1,0 @@
-export const displayChilds = (elementArray) => {
-  const root = document.getElementById("root");
-
-  elementArray.forEach((element) => root.appendChild(element));
-};
-
-export const clearChilds = () => {
-  const root = document.getElementById("root");
-
-  root.innerHTML = "";
-};
